Move inline module toggle into its own component
