+++ _inc/client/components/welcome-new-plan/premium.jsx
@@ -3,7 +3,6 @@
  */
 import PropTypes from 'prop-types';
 import React from 'react';
-import { connect } from 'react-redux';
 import { Component } from 'react';
 import { translate as __ } from 'i18n-calypso';
 import Card from 'components/card';
@@ -14,49 +13,7 @@ import analytics from 'lib/analytics';
  */
 import JetpackDialogue from 'components/jetpack-dialogue';
 import { imagePath } from 'constants/urls';
-
-// for inline settings
-import { ModuleSettingsForm as moduleSettingsForm } from 'components/module-settings/module-settings-form';
-import { getModule } from 'state/modules';
-import { ModuleToggle } from 'components/module-toggle';
-import SettingsGroup from 'components/settings-group';
-import decodeEntities from 'lib/decode-entities';
-
-const ModuleSettingsComponent = connect(
-	( state ) => {
-		return {
-			module: ( module_name ) => getModule( state, module_name ),
-		};
-	}
-)( moduleSettingsForm(
-	class extends Component {
-		render() {
-			// TODO: better class name
-			const module = this.props.module( this.props.module_slug );
-			return (
-				<div className="jp-upgrade-notice__enable-module">
-					<SettingsGroup
-						hasChild
-						disableInDevMode
-						module={ module }>
-
-						<ModuleToggle
-							slug={ this.props.module_slug }
-							disabled={ false }
-							activated={ this.props.getOptionValue( this.props.module_slug ) }
-							toggling={ this.props.isSavingAnyOption( this.props.module_slug ) }
-							toggleModule={ this.toggleModule }
-						>
-							<span className="jp-form-toggle-explanation">
-								{ decodeEntities( module.description ) }
-							</span>
-						</ModuleToggle>
-					</SettingsGroup>
-				</div>
-			);
-		}
-	}
-) );
+import InlineModuleToggle from 'components/module-settings/inline-module-toggle';
 
 class WelcomePremium extends Component {
 	constructor( props ) {
@@ -103,10 +60,22 @@ class WelcomePremium extends Component {
 						'following, you can grow your business with tools like payment buttons and ads.'
 					) }
 				</p>
+				<InlineModuleToggle module_slug="publicize" />
 				<img src={ imagePath + 'wordads.svg' } className="jp-welcome__svg" alt={ __( 'WordAds' ) } />
-				<ModuleSettingsComponent module_slug="wordads" />
+				<p>
+					{ __( 'Earn income by displaying high-quality ads on your site. ' +
+						'Jetpack Ads is a one-click feature designed to help you generate income from your WordPress site. ' +
+						'Jetpack Ads is powered by WordAds — the unique ad program from WordPress.com. '
+					) }
+				</p>
+				<InlineModuleToggle module_slug="wordads" />
 				<img src={ imagePath + 'wordads.svg' } className="jp-welcome__svg" alt={ __( 'VideoPress' ) } />
-				<ModuleSettingsComponent module_slug="videopress" />
+				<p>
+					{ __( 'VideoPress allows you to upload videos from your computer to be hosted on WordPress.com, ' +
+						'rather than on your host’s servers. You can then insert these on your self-hosted Jetpack site. '
+					) }
+				</p>
+				<InlineModuleToggle module_slug="videopress" />
 				<img src={ imagePath + 'security.svg' } className="jp-welcome__svg" alt={ __( 'Security' ) } />
 				<p>
 					{ __( 'Keeping your hard work safe is important, too. Jetpack Premium gives you brute force' +
